+++ src/Kiota.Builder/PathSegmenters/GoPathSegmenter.cs
@@ -1,4 +1,5 @@
-﻿using System.Collections.Generic;
+﻿using System;
+using System.Collections.Generic;
 using System.Linq;
 
 using Kiota.Builder.CodeDOM;
@@ -7,6 +8,67 @@ using Kiota.Builder.Extensions;
 namespace Kiota.Builder.PathSegmenters;
 public class GoPathSegmenter : CommonPathSegmenter
 {
+    private static readonly HashSet<string> specialFileNameSuffixes = new(StringComparer.OrdinalIgnoreCase) {
+        "test"       ,
+
+        "aix"       ,
+        "android"   ,
+        "darwin"    ,
+        "dragonfly" ,
+        "freebsd"   ,
+        "hurd"      ,
+        "illumos"   ,
+        "ios"       ,
+        "js"        ,
+        "linux"     ,
+        "nacl"      ,
+        "netbsd"    ,
+        "openbsd"   ,
+        "plan9"     ,
+        "solaris"   ,
+        "wasip1"    ,
+        "windows"   ,
+        "zos"       ,
+
+        "aix"       ,
+        "android"   ,
+        "darwin"    ,
+        "dragonfly" ,
+        "freebsd"   ,
+        "hurd"      ,
+        "illumos"   ,
+        "ios"       ,
+        "linux"     ,
+        "netbsd"    ,
+        "openbsd"   ,
+        "solaris"   ,
+
+        "386"         ,
+        "amd64"       ,
+        "amd64p32"    ,
+        "arm"         ,
+        "armbe"       ,
+        "arm64"       ,
+        "arm64be"     ,
+        "loong64"     ,
+        "mips"        ,
+        "mipsle"      ,
+        "mips64"      ,
+        "mips64le"    ,
+        "mips64p32"   ,
+        "mips64p32le" ,
+        "ppc"         ,
+        "ppc64"       ,
+        "ppc64le"     ,
+        "riscv"       ,
+        "riscv64"     ,
+        "s390"        ,
+        "s390x"       ,
+        "sparc"       ,
+        "sparc64"     ,
+        "wasm"        ,
+    };
+
     public GoPathSegmenter(string rootPath, string clientNamespaceName) : base(rootPath, clientNamespaceName) { }
     public override string FileSuffix => ".go";
     public override IEnumerable<string> GetAdditionalSegment(CodeElement currentElement, string fileName)
@@ -17,12 +79,13 @@ public class GoPathSegmenter : CommonPathSegmenter
             _ => Enumerable.Empty<string>(),
         };
     }
+
     public override string NormalizeFileName(CodeElement currentElement)
     {
         return currentElement switch
         {
             CodeNamespace => "go",
-            _ => GetLastFileNameSegment(currentElement).ToSnakeCase().ShortenFileName(),
+            _ => GetLastFileNameSegment(currentElement).ToSnakeCase().EscapeSuffix(specialFileNameSuffixes).ShortenFileName(),
         };
     }
     public override string NormalizeNamespaceSegment(string segmentName) => segmentName?.ToLowerInvariant() ?? string.Empty;