+++ CHANGELOG.md
@@ -9,10 +9,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Suppress CS1591 when generating CSharp code and documentation is not available
+- Added file name suffix escaping in Go to avoid generating files with reserved suffixes. [#4407](https://github.com/microsoft/kiota/issues/4407)
+
 ### Changed
 
+- Fixed a bug where TypeScript deserialization would fail on Uppercase properties.[#4479](https://github.com/microsoft/kiota/issues/4479)
 - Changed URI template generation to reuse templates when required templates are absent across operations.
-- Updated reserved name providers for Java and Php so that "object" can be escaped. 
+- Updated reserved name providers for Java and Php so that "object" can be escaped.
+- Fixes request builder disambiguation when child nodes had different prefixes for different paths with same parameters.[#4448](https://github.com/microsoft/kiota/issues/4448) 
+- Do not generate CS8603 warnings when enabling backing store in CSharp generation.
 
 ## [1.13.0] - 2024-04-04